Refactor dataset into its own module
